Add files via upload
The PDF file "Recommended Food Storage Times Cold and Dry Refrigerated and Frozen Foods" is a dataset intended for the Food Storage section of our website. The PDF file contains multiple tables that list the proper storage times and conditions for different food products. 

The code is written to extract the tables from the PDF file and save it in CSV format. The tables are extracted one by one from the PDF. Then, cleaned individually and combined into one table to create the final CSV file.